Extract `Curve` out of `PrimitiveCurve`
