+++ app/src/commonMain/kotlin/diy/lingerie/geometry/curves/Curve.kt
@@ -1,0 +1,54 @@
+package diy.lingerie.geometry.curves
+
+import diy.lingerie.algebra.NumericObject
+import diy.lingerie.algebra.equalsWithTolerance
+import diy.lingerie.geometry.Point
+import kotlin.jvm.JvmInline
+
+abstract class Curve : NumericObject {
+    @JvmInline
+    value class Coord(
+        /**
+         * The t-value for the basis function
+         */
+        val t: Double,
+    ) : NumericObject {
+        val complement: Coord
+            get() = Coord(
+                t = 1.0 - t,
+            )
+
+        init {
+            require(t in 0.0..1.0)
+        }
+
+        override fun equalsWithTolerance(
+            other: NumericObject,
+            tolerance: NumericObject.Tolerance,
+        ): Boolean = when {
+            other !is Coord -> false
+            !t.equalsWithTolerance(other.t, tolerance) -> false
+            else -> true
+        }
+    }
+
+    /**
+     * The intersection of two curves
+     */
+    abstract class Intersection {
+        /**
+         * The point of intersection
+         */
+        abstract val point: Point
+
+        /**
+         * The coordinate of the intersection point on the subject curve
+         */
+        abstract val coord: Coord
+
+        /**
+         * The coordinate of the intersection point on the other curve
+         */
+        abstract val otherCoord: Coord
+    }
+}
