+++ app/src/commonMain/kotlin/diy/lingerie/geometry/curves/PrimitiveCurve.kt
@@ -1,58 +1,10 @@
 package diy.lingerie.geometry.curves
 
 import diy.lingerie.algebra.NumericObject
-import diy.lingerie.algebra.equalsWithTolerance
 import diy.lingerie.geometry.Point
 import diy.lingerie.geometry.transformations.Transformation
-import kotlin.jvm.JvmInline
 
-abstract class PrimitiveCurve : NumericObject {
-    @JvmInline
-    value class Coord(
-        /**
-         * The t-value for the basis function
-         */
-        val t: Double,
-    ) : NumericObject {
-        val complement: Coord
-            get() = Coord(
-                t = 1.0 - t,
-            )
-
-        init {
-            require(t in 0.0..1.0)
-        }
-
-        override fun equalsWithTolerance(
-            other: NumericObject,
-            tolerance: NumericObject.Tolerance,
-        ): Boolean = when {
-            other !is Coord -> false
-            !t.equalsWithTolerance(other.t, tolerance) -> false
-            else -> true
-        }
-    }
-
-    /**
-     * The intersection of two curves
-     */
-    abstract class Intersection {
-        /**
-         * The point of intersection
-         */
-        abstract val point: Point
-
-        /**
-         * The coordinate of the intersection point on the subject curve
-         */
-        abstract val coord: Coord
-
-        /**
-         * The coordinate of the intersection point on the other curve
-         */
-        abstract val otherCoord: Coord
-    }
-
+abstract class PrimitiveCurve : Curve() {
     /**
      * Find the intersections of this curve (also referred to as the "subject
      * curve") with the [other] curve.
